#42: Created an integration test using `test-db-builder-java`.
