+++ .settings/org.eclipse.jdt.ui.prefs
@@ -34,6 +34,7 @@ sp_cleanup.never_use_blocks=false
 sp_cleanup.never_use_parentheses_in_expressions=false
 sp_cleanup.on_save_use_additional_actions=true
 sp_cleanup.organize_imports=true
+sp_cleanup.push_down_negation=false
 sp_cleanup.qualify_static_field_accesses_with_declaring_class=false
 sp_cleanup.qualify_static_member_accesses_through_instances_with_declaring_class=true
 sp_cleanup.qualify_static_member_accesses_through_subtypes_with_declaring_class=true
@@ -46,6 +47,7 @@ sp_cleanup.remove_redundant_type_arguments=true
 sp_cleanup.remove_trailing_whitespaces=true
 sp_cleanup.remove_trailing_whitespaces_all=true
 sp_cleanup.remove_trailing_whitespaces_ignore_empty=false
+sp_cleanup.remove_unnecessary_array_creation=false
 sp_cleanup.remove_unnecessary_casts=true
 sp_cleanup.remove_unnecessary_nls_tags=true
 sp_cleanup.remove_unused_imports=true
