+++ pom.xml
@@ -40,7 +40,7 @@
         <dependency>
             <groupId>com.exasol</groupId>
             <artifactId>exasol-virtual-schema</artifactId>
-            <version>1.0.0</version>
+            <version>2.0.0</version>
         </dependency>
         <dependency>
             <groupId>com.exasol</groupId>
@@ -117,6 +117,12 @@
             <artifactId>slf4j-jdk14</artifactId>
             <version>1.7.30</version>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+           <groupId>com.exasol</groupId>
+           <artifactId>test-db-builder-java</artifactId>
+           <version>0.1.0</version>
+           <scope>test</scope>
         </dependency>
     </dependencies>
     <build>
