+++ src/test/java/com/exasol/adapter/dialects/rls/TableProtectionStatusReaderTest.java
@@ -1,41 +1,70 @@
 package com.exasol.adapter.dialects.rls;
 
+import static com.exasol.adapter.dialects.rls.RowLevelSecurityDialectConstants.*;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import java.sql.*;
 
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 public class TableProtectionStatusReaderTest {
-    @Test
-    void testIsTableSecuredWithRolesTrue() throws SQLException {
-        final DatabaseMetaData metadataMock = Mockito.mock(DatabaseMetaData.class);
-        final ResultSet resultSet = Mockito.mock(ResultSet.class);
-        when(metadataMock.getColumns(any(), any(), any(), any())).thenReturn(resultSet);
+    private static final String PROTECTED_TABLE_NAME = "TableA";
+
+    @CsvSource({ //
+            EXA_ROW_ROLES_COLUMN_NAME + ", true, false, false", //
+            EXA_ROW_TENANT_COLUMN_NAME + ", false, true, false", //
+            EXA_ROW_GROUP_COLUMN_NAME + ", false, false, true" //
+    })
+    @ParameterizedTest
+    void testIsTableXProteced(final String protectionColumnName, final boolean roleProtected,
+            final boolean tenantProtected, final boolean groupProtected) throws SQLException {
+        final ResultSet resultSet = mockColumnMetadataWithProtectionColumn(protectionColumnName);
+        final TableProtectionStatus status = readProtectionStatusFromMockedMetadata(resultSet);
+        assertAll(
+                () -> assertThat("role-protected", status.isTableRoleProtected(PROTECTED_TABLE_NAME),
+                        equalTo(roleProtected)),
+                () -> assertThat("tenant-protected", status.isTableTenantProtected(PROTECTED_TABLE_NAME),
+                        equalTo(tenantProtected)),
+                () -> assertThat("group-protected", status.isTableGroupProtected(PROTECTED_TABLE_NAME),
+                        equalTo(groupProtected)));
+    }
+
+    private ResultSet mockColumnMetadataWithProtectionColumn(final String protectionColumnName) throws SQLException {
+        final ResultSet resultSet = mock(ResultSet.class);
         when(resultSet.next()).thenReturn(true, false);
-        final String tableName = "TableA";
-        when(resultSet.getString(3)).thenReturn(tableName);
-        when(resultSet.getString(4)).thenReturn(RowLevelSecurityDialectConstants.EXA_ROW_ROLES_COLUMN_NAME);
+        when(resultSet.getString(3)).thenReturn(PROTECTED_TABLE_NAME);
+        when(resultSet.getString(4)).thenReturn(protectionColumnName);
+        return resultSet;
+    }
+
+    private TableProtectionStatus readProtectionStatusFromMockedMetadata(final ResultSet resultSet)
+            throws SQLException {
+        final DatabaseMetaData metadataMock = mock(DatabaseMetaData.class);
+        when(metadataMock.getColumns(any(), any(), any(), any())).thenReturn(resultSet);
         final TableProtectionStatus tableProtectionStatus = new TableProtectionStatusReader(metadataMock)
                 .read("IrrelevantCatalog", "IrrelevantSchema");
-        assertAll(() -> assertThat(tableProtectionStatus.isTableRoleProtected(tableName), equalTo(true)),
-                () -> assertThat(tableProtectionStatus.isTableTenantProtected(tableName), equalTo(false)));
+        return tableProtectionStatus;
     }
 
     @Test
     void testIsTableSecuredWithRolesFalse() throws SQLException {
-        final DatabaseMetaData metadataMock = Mockito.mock(DatabaseMetaData.class);
-        final ResultSet resultSet = Mockito.mock(ResultSet.class);
-        when(metadataMock.getColumns(any(), any(), any(), any())).thenReturn(resultSet);
+        final ResultSet resultSet = mockColumnMetadataForNoProtectedTables();
+        final TableProtectionStatus status = readProtectionStatusFromMockedMetadata(resultSet);
+        assertAll(() -> assertThat("role-protected", status.isTableRoleProtected(""), equalTo(false)),
+                () -> assertThat("tenant-protected", status.isTableTenantProtected(""), equalTo(false)),
+                () -> assertThat("group-protected", status.isTableGroupProtected(""), equalTo(false)));
+    }
+
+    private ResultSet mockColumnMetadataForNoProtectedTables() throws SQLException {
+        final ResultSet resultSet = mock(ResultSet.class);
         when(resultSet.next()).thenReturn(false);
-        final TableProtectionStatus tableProtectionStatus = new TableProtectionStatusReader(metadataMock)
-                .read("IrrelevantCatalog", "IrrelevantSchema");
-        assertAll(() -> assertThat(tableProtectionStatus.isTableRoleProtected(""), equalTo(false)),
-                () -> assertThat(tableProtectionStatus.isTableTenantProtected(""), equalTo(false)));
+        return resultSet;
     }
 }
